refactor(timeline): move dummy events data to separate file
